// frontend2/src/utils/types/model/StatusBccEnum.ts
/**
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import * as models from './models';

export enum StatusBccEnum {
    New = <any> 'NEW',
    Que = <any> 'QUE',
    Run = <any> 'RUN',
    Try = <any> 'TRY',
    Ok = <any> 'OK!',
    Err = <any> 'ERR',
    Can = <any> 'CAN'
}